requirements.txt
aiohttp==3.7.4.post0
anyio==2.2.0
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
backcall==0.2.0
bleach==3.3.0
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
decorator==5.0.7
defusedxml==0.7.1
deprecation==2.1.0
Django==3.2
entrypoints==0.3
hyperlink==21.0.0
idna==3.1
ipykernel==5.5.3
ipython==7.23.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.2
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
Send2Trash==1.5.0
simpervisor==0.4
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
terminado==0.9.4
testpath==0.4.4
tomlkit==0.7.0
tornado==6.1
traitlets==5.0.5
txaio==21.2.1
typing-extensions==3.10.0.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
yarl==1.6.3
gunicorn==20.1.0
whitenoise==5.2.0
psycopg2 == 2.8.6
python-dotenv==0.17.1
imdbpy==2021.4.18
